# docs/api-contract.yaml
openapi: 3.0.3
info:
  title: Semantic Business Map API
  version: 1.0.0
  description: Evidence-first application contract. Source for generated frontend types. No graph snippets are returned by the canvas endpoint.
servers:
  - url: /api/v1
paths:
  /projects:
    get:
      operationId: listProjects
      responses:
        '200':
          description: Projects visible to the current user
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Project' } }
    post:
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectInput' }
      responses:
        '201':
          description: Created project
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
  /projects/{projectId}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    patch:
      operationId: updateProject
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectInput' }
      responses:
        '200':
          description: Updated project
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
    delete:
      operationId: deleteProject
      responses:
        '204': { description: Project archived; immutable history and audit are preserved }
...
  /analysis-runs/{analysisRunId}/canvas:
    parameters:
      - $ref: '#/components/parameters/RunId'
    get:
      operationId: getCanvas
      parameters:
        - in: query
          name: view
          schema: { $ref: '#/components/schemas/CanvasView' }
        - in: query
          name: depth
          schema: { type: integer, minimum: 1, maximum: 5, default: 2 }
        - in: query
          name: levelOfDetail
          schema: { type: integer, minimum: 0, maximum: 3, default: 2 }
        - in: query
          name: onlyChanged
          schema: { type: boolean, default: false }
        - in: query
          name: minConfidence
          schema: { type: number, minimum: 0, maximum: 1, default: 0 }
        - in: query
          name: rootNodeId
          schema: { type: string }
        - in: query
          name: search
          schema: { type: string }
      responses:
        '200':
          description: Bounded projection with no source snippets
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CanvasProjection' }
  /analysis-runs/{analysisRunId}/layout:
    parameters:
      - $ref: '#/components/parameters/RunId'
      - in: query
        name: view
        required: true
        schema: { $ref: '#/components/schemas/CanvasView' }
    get:
      operationId: getLayout
      responses:
        '200':
          description: Stable key positions and viewport; new layouts return empty positions
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CanvasLayout' }
        '404': { description: No saved layout }
    put:
      operationId: saveLayout
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CanvasLayout' }
      responses:
        '200': { description: Layout saved }
        '204': { description: Layout saved }
  /semantic/nodes/{nodeId}:
    parameters:
      - $ref: '#/components/parameters/NodeId'
    get:
      operationId: getSemanticNode
      responses:
        '200':
          description: Node and separate interpretation layers
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SemanticNodeDetail' }
  /semantic/nodes/{nodeId}/evidence:
    parameters:
      - $ref: '#/components/parameters/NodeId'
    get:
      operationId: getNodeEvidence
      responses:
        '200':
          description: Evidence with exact source locations
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Evidence' } }
  /semantic/edges/{edgeId}/evidence:
    parameters:
      - name: edgeId
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: getEdgeEvidence
      responses:
        '200':
          description: Evidence explaining the selected relationship
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Evidence' } }
...
  /projects/{projectId}/review-queue:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      operationId: getReviewQueue
      responses:
        '200':
          description: Nodes awaiting review
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/SemanticNode' } }
  /projects/{projectId}/semantic-diff:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      operationId: getSemanticDiff
      parameters:
        - in: query
          name: fromAnalysisRun
          required: true
          schema: { type: string }
        - in: query
          name: toAnalysisRun
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Deterministic semantic changes between two runs
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/SemanticChange' } }
...
  /analysis-runs/{analysisRunId}/quarantined-facts:
    parameters:
      - $ref: '#/components/parameters/RunId'
    get:
      operationId: getQuarantinedFacts
      responses:
        '200':
          $ref: '#/components/responses/Inspection'
  /analysis-runs/{analysisRunId}/unresolved-symbols:
    parameters:
      - $ref: '#/components/parameters/RunId'
    get:
      operationId: getUnresolvedSymbols
      responses:
        '200':
          $ref: '#/components/responses/Inspection'
...
components:
  parameters:
    ProjectId:
      in: path
      name: projectId
      required: true
      schema: { type: string }
    RunId:
      in: path
      name: analysisRunId
      required: true
      schema: { type: string }
    NodeId:
      in: path
      name: nodeId
      required: true
      schema: { type: string }
  responses:
    Inspection:
      description: Structured diagnostic data; fields vary by analyzer capability
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InspectionData'
...
  schemas:
    ProjectInput:
      type: object
      required: [name, description, repositoryPath]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        description: { type: string, maxLength: 2000 }
        repositoryPath: { type: string, maxLength: 2000 }
    Project:
      allOf:
        - $ref: '#/components/schemas/ProjectInput'
        - type: object
          required: [id, createdAt]
          properties:
            id: { type: string }
            createdAt: { type: string }
...
    CanvasView:
      type: string
      enum: [BUSINESS, DEVELOPER, DATA_OWNERSHIP, SECURITY, DATA_PIPELINE, CONFIDENCE]
    SemanticNode:
      type: object
      required: [id, stableKey, kind, label, subtitle, confidence, supportLevel, evidenceCount, reviewStatus, badges, properties]
      properties:
        id: { type: string }
        stableKey: { type: string }
        kind: { type: string }
        label: { type: string }
        subtitle: { type: string, nullable: true }
        confidence: { type: number, minimum: 0, maximum: 1 }
        supportLevel: { type: string }
        evidenceCount: { type: integer, minimum: 0 }
        reviewStatus: { type: string }
        badges: { type: array, items: { type: string } }
        properties: { type: object, additionalProperties: {} }
    SemanticNodeDetail:
      allOf:
        - $ref: '#/components/schemas/SemanticNode'
        - type: object
          properties:
            facts: { type: array, items: {} }
            scenarios: { type: array, items: {} }
            assertions: { type: array, items: {} }
            enrichments: { type: array, items: {} }
            reviews: { type: array, items: {} }
    SemanticEdge:
      type: object
      required: [id, source, target, kind, label, confidence, evidenceCount]
      properties:
        id: { type: string }
        source: { type: string }
        target: { type: string }
        kind: { type: string }
        label: { type: string, nullable: true }
        confidence: { type: number, minimum: 0, maximum: 1 }
        evidenceCount: { type: integer, minimum: 0 }
    CanvasProjection:
      type: object
      required: [nodes, edges, truncated, totalNodes]
      properties:
        nodes: { type: array, items: { $ref: '#/components/schemas/SemanticNode' } }
        edges: { type: array, items: { $ref: '#/components/schemas/SemanticEdge' } }
        truncated: { type: boolean }
        totalNodes: { type: integer, minimum: 0 }
    CanvasLayout:
      type: object
      required: [positions, viewport]
      properties:
        positions:
          type: object
          additionalProperties:
            type: object
            required: [x, y]
            properties:
              x: { type: number }
              y: { type: number }
        viewport:
          type: object
          required: [x, y, zoom]
          properties:
            x: { type: number }
            y: { type: number }
            zoom: { type: number, minimum: 0.01, maximum: 10 }
    Evidence:
      type: object
      required: [id, filePath, startLine, endLine, snippet, snippetHash, analyzerId, analyzerVersion, imageDigest, origin, verified]
      properties:
        id: { type: string }
        filePath: { type: string }
        startLine: { type: integer, minimum: 1 }
        endLine: { type: integer, minimum: 1 }
        snippet: { type: string, nullable: true }
        snippetHash: { type: string, nullable: true }
        analyzerId: { type: string }
        analyzerVersion: { type: string, nullable: true }
        imageDigest: { type: string, nullable: true }
        origin: { type: string }
        verified: { type: boolean }
        sourceAvailable:
          type: boolean
          description: False when retention has removed the source snippet; evidence metadata remains available.
...
    SemanticChange:
      type: object
      required: [id, changeType, subjectStableKey, impactType, before, after, beforeEvidenceIds, afterEvidenceIds, confidence]
      properties:
        id: { type: string }
        changeType: { type: string }
        subjectStableKey: { type: string }
        impactType: { type: string }
        before: {}
        after: {}
        beforeEvidenceIds: { type: array, items: { type: string } }
        afterEvidenceIds: { type: array, items: { type: string } }
        confidence: { type: number, minimum: 0, maximum: 1 }
...
    InspectionData:
      oneOf:
        - { type: object, additionalProperties: {} }
        - { type: array, items: {} }
...